Extract shared app action builders
