+++ app/src/main/java/com/d4rk/android/apps/apptoolkit/app/apps/favorites/ui/FavoriteAppsScreen.kt
@@ -6,7 +6,8 @@ import androidx.compose.material.icons.outlined.Android
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
+import com.d4rk.android.apps.apptoolkit.app.apps.common.buildOnAppClick
+import com.d4rk.android.apps.apptoolkit.app.apps.common.buildOnShareClick
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.d4rk.android.apps.apptoolkit.R
@@ -20,9 +21,6 @@ import com.d4rk.android.libs.apptoolkit.core.domain.model.ui.UiStateScreen
 import com.d4rk.android.libs.apptoolkit.core.ui.components.ads.rememberAdsEnabled
 import com.d4rk.android.libs.apptoolkit.core.ui.components.layouts.NoDataScreen
 import com.d4rk.android.libs.apptoolkit.core.ui.components.layouts.ScreenStateHandler
-import com.d4rk.android.libs.apptoolkit.core.utils.helpers.AppInfoHelper
-import com.d4rk.android.libs.apptoolkit.core.utils.helpers.IntentsHelper
-import kotlinx.coroutines.launch
 import org.koin.compose.koinInject
 import org.koin.compose.viewmodel.koinViewModel
 
@@ -36,32 +34,8 @@ fun FavoriteAppsRoute(paddingValues: PaddingValues) {
     val onFavoriteToggle: (String) -> Unit = remember(viewModel) { { pkg -> viewModel.toggleFavorite(pkg) } }
     val onRetry: () -> Unit = remember(viewModel) { { viewModel.onEvent(FavoriteAppsEvent.LoadFavorites) } }
     val dispatchers: DispatcherProvider = koinInject()
-    val appInfoHelper = remember { AppInfoHelper(dispatchers) }
-    val coroutineScope = rememberCoroutineScope()
-    val onAppClick: (AppInfo) -> Unit = remember(context, coroutineScope) {
-        { appInfo ->
-            coroutineScope.launch {
-                if (appInfo.packageName.isNotEmpty()) {
-                    if (appInfoHelper.isAppInstalled(context, appInfo.packageName)) {
-                        if (!appInfoHelper.openApp(context, appInfo.packageName)) {
-                            IntentsHelper.openPlayStoreForApp(context, appInfo.packageName)
-                        }
-                    } else {
-                        IntentsHelper.openPlayStoreForApp(context, appInfo.packageName)
-                    }
-                }
-            }
-        }
-    }
-    val onShareClick: (AppInfo) -> Unit = remember(context) {
-        {
-            IntentsHelper.shareApp(
-                context = context,
-                shareMessageFormat = com.d4rk.android.libs.apptoolkit.R.string.summary_share_message,
-                packageName = it.packageName
-            )
-        }
-    }
+    val onAppClick: (AppInfo) -> Unit = buildOnAppClick(dispatchers, context)
+    val onShareClick: (AppInfo) -> Unit = buildOnShareClick(context)
 
     FavoriteAppsScreen(
         screenState = screenState,
